requirements.txt
fastapi>=0.68.0
uvicorn>=0.13.0
sqlalchemy>=2.0.0
aiosqlite>=0.17.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
anyio>=3.0.0
pylint==3.3.6
